+++ app/src/main/java/com/github/mobile/ui/comment/RawCommentFragment.java
@@ -16,6 +16,7 @@
 package com.github.mobile.ui.comment;
 
 import android.app.Activity;
+import android.os.Build;
 import android.os.Bundle;
 import android.text.Editable;
 import android.view.LayoutInflater;
@@ -50,7 +51,7 @@ public class RawCommentFragment extends DialogFragment {
             @Override
             public void afterTextChanged(Editable s) {
                 Activity activity = getActivity();
-                if (activity != null)
+                if (activity != null && Build.VERSION.SDK_INT >= Build.VERSION_CODES.HONEYCOMB)
                     activity.invalidateOptionsMenu();
             }
         });